refactor(launcher): Move avs FS dir creation stuff to module
Part of the entire "setup AVS stuff", so it fits well into
the avs_context_init method which declutters the launcher
main method significantly
