+++ src/main/launcher/main.c
@@ -179,75 +179,6 @@ static void log_avs_fs_dir(const char *path)
     avs_fs_closedir(dir);
 }
 
-static void create_avs_config_fs_dir(
-        struct property *prop,
-        struct property_node *node,
-        const char *folder_name)
-{
-    char fs_path[1024];
-    char fs_type[255];
-    char device_path[1024];
-    struct property_node *fs_node;
-    int res;
-
-    memset(fs_path, 0, sizeof(fs_path));
-    memset(fs_type, 0, sizeof(fs_type));
-
-    str_cpy(fs_path, sizeof(fs_path), "/fs/");
-    str_cat(fs_path, sizeof(fs_path), folder_name);
-
-    fs_node = property_search(prop, node, fs_path);
-
-    if (!fs_node) {
-        log_warning("Could not find file system node %s in avs configuration", fs_path);
-        return;
-    }
-
-    res = property_node_refer(prop, fs_node, "device", PROPERTY_TYPE_STR,
-        device_path, sizeof(device_path));
-
-    if (res < 0) {
-        log_fatal("Getting 'device' attribute from avs config entry %s failed", fs_path);
-    }
-
-    // 'fstype' attribute is optional and defaults to value 'fs'
-    if (!property_search(prop, fs_node, "fstype")) {
-        if (path_exists(device_path)) {
-            // skip if exists already
-            return;
-        }
-
-        log_misc("Creating avs directory %s", device_path);
-
-        if (!path_mkdir(device_path)) {
-            log_fatal("Creating directory %s failed", device_path);
-        }
-    } else {
-        res = property_node_refer(prop, fs_node, "fstype", PROPERTY_TYPE_STR,
-                fs_type, sizeof(fs_type));
-
-        if (res < 0) {
-            log_fatal("Getting 'fstype' attribute from avs config entry %s failed", fs_path);
-        }
-
-        if (!strcmp(fs_type, "fs") || !strcmp(fs_type, "nvram")) {
-            if (path_exists(device_path)) {
-                // skip if exists already
-                return;
-            }
-
-            log_misc("Creating avs directory %s", device_path);
-
-            if (!path_mkdir(device_path)) {
-                log_fatal("Creating directory %s failed", device_path);
-            }
-        } else {
-            log_fatal("Cannot create folders for unsupported file system type %s of path %s in avs config",
-                fs_type, fs_path);
-        }
-    }
-}
-
 static void log_property_tree(struct property *property)
 {
     log_property_node_tree_rec(property_search(property, NULL, "/"), "");
@@ -377,18 +308,13 @@ int main(int argc, const char **argv)
 
     bootstrap_config_update_avs(&bs, avs_config_root);
 
-    log_misc("Creating file system directories for nvram and raw if not exist...");
-
-    // create nvram and raw directories if possible for non-mounttable configurations
-    create_avs_config_fs_dir(avs_config, avs_config_root, "nvram");
-    create_avs_config_fs_dir(avs_config, avs_config_root, "raw");
-
     log_misc("Loading before hook dlls...");
     load_hook_dlls(&options.before_hook_dlls);
 
     log_misc("Initializing AVS...");
 
     avs_context_init(
+        avs_config,
         avs_config_root,
         options.avs_heap_size,
         options.std_heap_size,
